Refactor HeroController into partial classes
