+++ Assets/Scripts/Hero/HeroController.cs
@@ -35,7 +35,7 @@ namespace TimelessEchoes.Hero
     /// Exposes computed properties like Damage, AttackRate, MoveSpeed, Defense, and MaxHealthValue
     /// which include permanent upgrades and active buff multipliers.
     /// </summary>
-    public class HeroController : MonoBehaviour
+    public partial class HeroController : MonoBehaviour
     {
         public static HeroController Instance { get; private set; }
         private static bool nextIsEcho;
@@ -164,44 +164,6 @@ namespace TimelessEchoes.Hero
         public Animator Animator => animator;
         public bool InCombat => state == State.Combat;
 
-        /// <summary>
-        ///     Current attack damage after upgrades, buffs and dice multipliers.
-        /// </summary>
-        public float Damage =>
-            (baseDamage + damageBonus + gearDamageBonus) *
-            (buffController != null ? buffController.DamageMultiplier : 1f) *
-            combatDamageMultiplier;
-
-        /// <summary>
-        ///     Base attack damage after permanent upgrades and gear (before buffs and dice multipliers).
-        /// </summary>
-        public float BaseDamage => baseDamage + damageBonus + gearDamageBonus;
-
-        /// <summary>
-        ///     Current attacks per second after upgrades and buffs.
-        /// </summary>
-        public float AttackRate => CurrentAttackRate;
-
-        /// <summary>
-        ///     Current movement speed after upgrades and buffs.
-        /// </summary>
-        public float MoveSpeed =>
-            (baseMoveSpeed + moveSpeedBonus + gearMoveSpeedBonus) *
-            (buffController != null ? buffController.MoveSpeedMultiplier : 1f);
-
-        /// <summary>
-        ///     Maximum health after upgrades.
-        /// </summary>
-        public float MaxHealthValue => baseHealth + healthBonus + gearHealthBonus;
-
-        private float CurrentAttackRate =>
-            (baseAttackSpeed + attackSpeedBonus + gearAttackSpeedBonus) *
-            (buffController != null ? buffController.AttackSpeedMultiplier : 1f);
-
-        public float Defense =>
-            (baseDefense + defenseBonus + gearDefenseBonus) *
-            (buffController != null ? buffController.DefenseMultiplier : 1f);
-
         private void Awake()
         {
             if (nextIsEcho)
@@ -317,38 +279,6 @@ namespace TimelessEchoes.Hero
                             }, gm.ReaperSpawnOffset);
                         ReaperSpawnedByDistance = true;
                     }
-                }
-            }
-        }
-
-        private void RecalculateGearBonuses()
-        {
-            var equip = TimelessEchoes.Gear.EquipmentController.Instance ??
-                        FindFirstObjectByType<TimelessEchoes.Gear.EquipmentController>();
-            if (equip == null)
-            {
-                gearDamageBonus = gearAttackSpeedBonus = gearDefenseBonus = gearHealthBonus = gearMoveSpeedBonus = 0f;
-                return;
-            }
-
-            gearDamageBonus = equip.GetTotalForMapping(TimelessEchoes.Gear.HeroStatMapping.Damage);
-            gearAttackSpeedBonus = equip.GetTotalForMapping(TimelessEchoes.Gear.HeroStatMapping.AttackRate);
-            gearDefenseBonus = equip.GetTotalForMapping(TimelessEchoes.Gear.HeroStatMapping.Defense);
-            gearHealthBonus = equip.GetTotalForMapping(TimelessEchoes.Gear.HeroStatMapping.MaxHealth);
-            gearMoveSpeedBonus = equip.GetTotalForMapping(TimelessEchoes.Gear.HeroStatMapping.MoveSpeed);
-
-            // If MaxHealth changes, re-init health so UI reflects new max
-            if (health != null)
-            {
-                var oldMax = Mathf.RoundToInt(health.MaxHealth);
-                var current = Mathf.RoundToInt(health.CurrentHealth);
-                var newMax = Mathf.RoundToInt(baseHealth + healthBonus + gearHealthBonus);
-                if (Mathf.Abs(newMax - oldMax) > 0.01f && newMax > 0)
-                {
-                    var newCurrent = Mathf.Min(current + (newMax - oldMax), newMax);
-                    health.Init(newMax);
-                    if (newCurrent < newMax)
-                        health.TakeDamage(newMax - newCurrent);
                 }
             }
         }
@@ -481,644 +411,12 @@ namespace TimelessEchoes.Hero
                 equip.OnEquipmentChanged -= RecalculateGearBonuses;
         }
 
-        private void OnMilestoneUnlocked(Skill skill, MilestoneBonus milestone)
-        {
-            if (milestone != null && milestone.type == MilestoneType.StatIncrease)
-            {
-                var oldMax = health != null ? health.MaxHealth : 0f;
-                var oldCurrent = health != null ? health.CurrentHealth : 0f;
-                ApplyStatUpgrades();
-
-                if (health != null)
-                {
-                    var newMax = Mathf.RoundToInt(baseHealth + healthBonus);
-                    if (newMax > 0 && Mathf.Abs(newMax - oldMax) > 0.01f)
-                    {
-                        var newCurrent = Mathf.Min(oldCurrent + (newMax - oldMax), newMax);
-                        health?.Init(newMax);
-                        if (newCurrent < newMax && health != null)
-                            health.TakeDamage(newMax - newCurrent);
-                    }
-                }
-            }
-        }
-
-        private void ApplyStatUpgrades()
-        {
-            var controller = StatUpgradeController.Instance;
-            if (controller == null)
-                Log("StatUpgradeController missing", TELogCategory.Upgrade, this);
-            var skillController = SkillController.Instance;
-            if (skillController == null)
-                Log("SkillController missing", TELogCategory.Upgrade, this);
-            if (controller == null) return;
-
-            foreach (var upgrade in controller.AllUpgrades)
-            {
-                if (upgrade == null) continue;
-                var baseVal = controller.GetBaseValue(upgrade);
-                var levelIncrease = UpgradeFeatureToggle.DisableStatUpgrades ? 0f : controller.GetIncrease(upgrade);
-                var flatBonus = skillController ? skillController.GetFlatStatBonus(upgrade) : 0f;
-                var percentBonus = skillController ? skillController.GetPercentStatBonus(upgrade) : 0f;
-
-                var totalBeforePercent = baseVal + levelIncrease + flatBonus;
-                var finalValue = totalBeforePercent * (1f + percentBonus);
-                var increase = finalValue - baseVal;
-                switch (upgrade.name)
-                {
-                    case "Health":
-                        baseHealth = baseVal;
-                        healthBonus = increase;
-                        break;
-                    case "Damage":
-                        baseDamage = baseVal;
-                        damageBonus = increase;
-                        break;
-                    case "Attack Rate":
-                        baseAttackSpeed = baseVal;
-                        attackSpeedBonus = increase;
-                        break;
-                    case "Move Speed":
-                        baseMoveSpeed = baseVal;
-                        moveSpeedBonus = increase;
-                        break;
-                    case "Defense":
-                        baseDefense = baseVal;
-                        defenseBonus = increase;
-                        break;
-                }
-            }
-        }
-
-        private void UpdateAnimation()
-        {
-            Vector2 vel = ai.desiredVelocity;
-            var dir = vel;
-
-            if (dir.sqrMagnitude < 0.0001f && setter != null && setter.target != null)
-                dir = setter.target.position - transform.position;
-
-            if (fourDirectional)
-            {
-                if (Mathf.Abs(dir.x) >= Mathf.Abs(dir.y))
-                    dir.y = 0f;
-                else
-                    dir.x = 0f;
-            }
-            else
-            {
-                dir.y = 0f;
-            }
-
-            if (dir.sqrMagnitude > 0.0001f)
-                lastMoveDir = dir;
-
-            animator.SetFloat("MoveX", lastMoveDir.x);
-            animator.SetFloat("MoveY", lastMoveDir.y);
-            animator.SetFloat("MoveMagnitude", vel.magnitude);
-
-            if (AutoBuffAnimator != null && AutoBuffAnimator.isActiveAndEnabled)
-            {
-                AutoBuffAnimator.SetFloat("MoveX", lastMoveDir.x);
-                AutoBuffAnimator.SetFloat("MoveY", lastMoveDir.y);
-                AutoBuffAnimator.SetFloat("MoveMagnitude", vel.magnitude);
-            }
-
-            if (spriteRenderer != null)
-                spriteRenderer.flipX = lastMoveDir.x < 0f;
-
-            if (autoBuffSpriteRenderer != null)
-                autoBuffSpriteRenderer.flipX = lastMoveDir.x < 0f;
-        }
-
-        public void SetTask(ITask task)
-        {
-            if (CurrentTask is BaseTask oldBase)
-                oldBase.ReleaseClaim(this);
-
-            Log($"Hero assigned task: {task?.GetType().Name ?? "None"}", TELogCategory.Task, this);
-            CurrentTask = task;
-            currentTaskName = task != null ? task.GetType().Name : "None";
-            currentTaskObject = task as MonoBehaviour;
-            state = State.Idle;
-
-            if (setter == null)
-                setter = GetComponent<AIDestinationSetter>();
-
-            if (setter != null)
-            {
-                setter.target = task?.Target;
-                if (ai != null)
-                    ai.Teleport(transform.position);
-                else
-                    ai?.SearchPath();
-            }
-
-            if (task is BaseTask newBase)
-                newBase.Claim(this);
-        }
-
-        /// <summary>
-        ///     Clear the reference to the active <see cref="TaskController" /> so
-        ///     this hero no longer receives task assignments.
-        /// </summary>
-        public void ClearTaskController()
-        {
-            taskController = null;
-        }
-
-        private void UpdateBehavior()
-        {
-            if (stats == null) return;
-
-            enemyRemovalBuffer.Clear();
-            foreach (var enemy in engagedEnemies)
-            {
-                var hp = enemy != null ? enemy.GetComponent<Health>() : null;
-                if (enemy == null || hp == null || hp.CurrentHealth <= 0f || !enemy.IsEngaged)
-                    enemyRemovalBuffer.Add(enemy);
-            }
-
-            foreach (var enemy in enemyRemovalBuffer)
-                UnregisterEngagedEnemy(enemy);
-
-            if (currentEnemy != null)
-            {
-                var hp = currentEnemy.GetComponent<Health>();
-                var enemyComp = currentEnemy.GetComponent<Enemy>();
-                if (hp == null || hp.CurrentHealth <= 0f || enemyComp == null || !engagedEnemies.Contains(enemyComp))
-                {
-                    currentEnemyHealth?.SetHealthBarVisible(false);
-                    currentEnemy = null;
-                    currentEnemyHealth = null;
-                }
-            }
-
-            Transform nearest = currentEnemy;
-            if (allowAttacks && nearest == null)
-            {
-                if (engagedEnemies.Count > 0)
-                {
-                    var best = float.PositiveInfinity;
-                    Enemy chosen = null;
-                    foreach (var enemy in engagedEnemies)
-                    {
-                        if (enemy == null) continue;
-
-                        Transform enemyTransform = null;
-                        try
-                        {
-                            enemyTransform = enemy.transform;
-                        }
-                        catch
-                        {
-                            // Enemy may have been destroyed mid-frame
-                            continue;
-                        }
-
-                        if (enemyTransform == null) continue;
-
-                        var dist = Vector2.Distance(transform.position, enemyTransform.position);
-                        if (dist < best)
-                        {
-                            best = dist;
-                            chosen = enemy;
-                        }
-                    }
-
-                    Transform chosenTransform = null;
-                    if (chosen != null)
-                    {
-                        try { chosenTransform = chosen.transform; } catch { chosenTransform = null; }
-                    }
-                    nearest = chosenTransform;
-                }
-                else
-                {
-                    var range = UnlimitedAggroRange ? combatAggroRange : stats.visionRange;
-                    nearest = FindNearestEnemy(range);
-                }
-            }
-
-            if (allowAttacks && nearest != null)
-            {
-                if (currentEnemy != nearest)
-                {
-                    currentEnemyHealth?.SetHealthBarVisible(false);
-                    currentEnemy = nearest;
-                    currentEnemyHealth = nearest.GetComponent<Health>();
-                    currentEnemyHealth?.SetHealthBarVisible(true);
-                }
-                else if (currentEnemyHealth == null)
-                {
-                    currentEnemyHealth = nearest.GetComponent<Health>();
-                    currentEnemyHealth?.SetHealthBarVisible(true);
-                }
-
-                if (state == State.PerformingTask && CurrentTask != null) CurrentTask.OnInterrupt(this);
-                HandleCombat(nearest);
-                return;
-            }
-
-            if (state == State.Combat && engagedEnemies.Count == 0)
-            {
-                Log("Hero exiting combat", TELogCategory.Combat, this);
-                combatDamageMultiplier = 1f;
-                isRolling = false;
-                diceRoller?.ResetRoll();
-                currentEnemyHealth?.SetHealthBarVisible(false);
-                currentEnemyHealth = null;
-                state = State.Idle;
-                taskController?.SelectEarliestTask(this);
-            }
-
-            if (CurrentTask == null || CurrentTask.IsComplete())
-            {
-                CurrentTask = null;
-                state = State.Idle;
-                taskController?.SelectEarliestTask(this);
-            }
-
-            if (CurrentTask == null)
-            {
-                var noVisibleTasks = taskController == null || !taskController.HasVisibleTasksForHero(this);
-                if (taskController == null || taskController.tasks.Count == 0 || (IsEcho && noVisibleTasks))
-                    AutoAdvance();
-                else
-                    setter.target = null;
-                return;
-            }
-
-            var dest = CurrentTask.Target;
-            if (setter.target != dest) setter.target = dest;
-
-            if (IsAtDestination(dest))
-            {
-                if (state != State.PerformingTask)
-                {
-                    state = State.PerformingTask;
-                    ai.canMove = !CurrentTask.BlocksMovement;
-                    CurrentTask.OnArrival(this);
-                }
-
-                CurrentTask.Tick(this);
-            }
-            else
-            {
-                state = State.MovingToTask;
-                ai.canMove = true;
-            }
-        }
-
-
-        private Transform FindNearestEnemy(float range)
-        {
-            Transform nearest = null;
-            var best = float.MaxValue;
-            var enemies = EnemyActivator.ActiveEnemies;
-            if (enemies == null)
-                return null;
-            Vector2 pos = transform.position;
-
-            var cam = EnemyActivator.Instance != null
-                ? EnemyActivator.Instance.GetComponent<Camera>()
-                : null;
-            Vector3 min = Vector3.zero, max = Vector3.zero;
-            var checkBounds = false;
-            if (cam != null)
-            {
-                const float padding = 2f;
-                min = cam.ViewportToWorldPoint(Vector3.zero) - Vector3.one * padding;
-                max = cam.ViewportToWorldPoint(Vector3.one) + Vector3.one * padding;
-                checkBounds = true;
-            }
-
-            foreach (var enemy in enemies)
-            {
-                if (enemy == null) continue;
-
-                Transform enemyTransform = null;
-                try
-                {
-                    enemyTransform = enemy.transform;
-                }
-                catch
-                {
-                    // Enemy may have been destroyed mid-iteration
-                    continue;
-                }
-
-                if (enemyTransform == null) continue;
-
-                if (checkBounds)
-                {
-                    var p = enemyTransform.position;
-                    if (p.x < min.x || p.x > max.x || p.y < min.y || p.y > max.y)
-                        continue;
-                }
-
-                var hp = enemy.GetComponent<Health>();
-                if (hp == null || hp.CurrentHealth <= 0f) continue;
-                var d = Vector2.Distance(pos, enemyTransform.position);
-                if (d <= range && d < best)
-                {
-                    best = d;
-                    nearest = enemyTransform;
-                }
-            }
-
-            return nearest;
-        }
-
-        private Transform FindNearestEnemy()
-        {
-            return FindNearestEnemy(stats.visionRange);
-        }
-
-        private void HandleCombat(Transform enemy)
-        {
-            ai.canMove = true;
-
-            if (state != State.Combat)
-            {
-                Log($"Hero entering combat with {enemy.name}", TELogCategory.Combat, this);
-                if (diceUnlocked && diceRoller != null && !isRolling)
-                {
-                    var rate = CurrentAttackRate;
-                    var cooldown = rate > 0f ? 1f / rate : 0.5f;
-                    StartCoroutine(RollForCombat(cooldown));
-                }
-            }
-
-            state = State.Combat;
-            setter.target = enemy;
-
-            var hp = enemy.GetComponent<Health>();
-            if (hp == null || hp.CurrentHealth <= 0f) return;
-
-            var dist = Vector2.Distance(transform.position, enemy.position);
-            if (dist <= stats.visionRange)
-            {
-                var rate = CurrentAttackRate;
-                var cooldown = rate > 0f ? 1f / rate : float.PositiveInfinity;
-                if (allowAttacks && Time.time - lastAttack >= cooldown && !isRolling)
-                {
-                    lastMoveDir = enemy.position - transform.position;
-                    Attack(enemy);
-                    lastAttack = Time.time;
-                }
-            }
-        }
-
-        private IEnumerator RollForCombat(float duration)
-        {
-            if (!diceUnlocked || diceRoller == null)
-                yield break;
-
-            isRolling = true;
-            lastAttack = Time.time;
-
-            yield return StartCoroutine(diceRoller.Roll(duration));
-
-            combatDamageMultiplier = 1f + 0.1f * diceRoller.Result;
-            isRolling = false;
-        }
-
-
-        private void OnEnemyEngage(Enemy enemy)
-        {
-            if (enemy == null)
-                return;
-
-            var hp = enemy.GetComponent<Health>();
-            if (hp == null || hp.CurrentHealth <= 0f)
-            {
-                UnregisterEngagedEnemy(enemy);
-                return;
-            }
-
-            if (enemy.IsEngaged)
-            {
-                if (!engagedEnemies.Contains(enemy))
-                {
-                    engagedEnemies.Add(enemy);
-
-                    Action deathHandler = () => UnregisterEngagedEnemy(enemy);
-                    hp.OnDeath += deathHandler;
-                    enemyDeathHandlers[enemy] = deathHandler;
-
-                    Action<Enemy> disengageHandler = null;
-                    disengageHandler = e =>
-                    {
-                        if (e == enemy && !e.IsEngaged)
-                            UnregisterEngagedEnemy(enemy);
-                    };
-                    Enemy.OnEngage += disengageHandler;
-                    enemyDisengageHandlers[enemy] = disengageHandler;
-                }
-            }
-            else
-            {
-                UnregisterEngagedEnemy(enemy);
-                return;
-            }
-
-            if (!allowAttacks)
-                return;
-
-            if (currentEnemy != null && currentEnemy != enemy.transform)
-                return;
-
-            if (currentEnemy == null)
-            {
-                currentEnemyHealth?.SetHealthBarVisible(false);
-                currentEnemy = enemy.transform;
-                currentEnemyHealth = hp;
-                currentEnemyHealth.SetHealthBarVisible(true);
-            }
-
-            if (state == State.PerformingTask && CurrentTask != null)
-                CurrentTask.OnInterrupt(this);
-
-            HandleCombat(enemy.transform);
-        }
-
-        private void UnregisterEngagedEnemy(Enemy enemy)
-        {
-            if (enemy == null)
-                return;
-
-            if (engagedEnemies.Remove(enemy))
-            {
-                if (enemyDeathHandlers.TryGetValue(enemy, out var death))
-                {
-                    var hp = enemy.GetComponent<Health>();
-                    if (hp != null)
-                        hp.OnDeath -= death;
-                    enemyDeathHandlers.Remove(enemy);
-                }
-
-                if (enemyDisengageHandlers.TryGetValue(enemy, out var disengage))
-                {
-                    Enemy.OnEngage -= disengage;
-                    enemyDisengageHandlers.Remove(enemy);
-                }
-            }
-
-            Transform enemyTransformSafe = null;
-            try
-            {
-                enemyTransformSafe = enemy.transform;
-            }
-            catch
-            {
-                enemyTransformSafe = null;
-            }
-
-            if (enemyTransformSafe != null && currentEnemy == enemyTransformSafe)
-            {
-                currentEnemyHealth?.SetHealthBarVisible(false);
-                currentEnemy = null;
-                currentEnemyHealth = null;
-            }
-        }
-
-        private void Attack(Transform target)
-        {
-            if (stats.projectilePrefab == null || target == null) return;
-
-            var enemy = target.GetComponent<Health>();
-            if (enemy == null || enemy.CurrentHealth <= 0f) return;
-
-            animator.Play("Attack");
-            if (AutoBuffAnimator != null && AutoBuffAnimator.isActiveAndEnabled)
-                AutoBuffAnimator.Play("Attack");
-
-            var origin = projectileOrigin ? projectileOrigin : transform;
-            var projObj = PoolManager.Get(stats.projectilePrefab);
-            projObj.transform.position = origin.position;
-            projObj.transform.rotation = Quaternion.identity;
-            var proj = projObj.GetComponent<Projectile>();
-            if (proj != null)
-            {
-                var killTracker = EnemyKillTracker.Instance;
-                if (killTracker == null)
-                    Log("EnemyKillTracker missing", TELogCategory.Combat, this);
-                var enemyStats = target.GetComponent<Enemy>()?.Stats;
-                var bonus = killTracker != null ? killTracker.GetDamageMultiplier(enemyStats) : 1f;
-                var dmgBase = (baseDamage + damageBonus + gearDamageBonus) *
-                              (buffController != null ? buffController.DamageMultiplier : 1f) *
-                              combatDamageMultiplier;
-                var total = dmgBase * bonus;
-
-                // Gear crit chance (2x damage). Only from gear.
-                var equip = TimelessEchoes.Gear.EquipmentController.Instance ??
-                            FindFirstObjectByType<TimelessEchoes.Gear.EquipmentController>();
-                float critChance = 0f;
-                if (equip != null)
-                {
-                    var crafting = TimelessEchoes.Gear.CraftingService.Instance ??
-                                   FindFirstObjectByType<TimelessEchoes.Gear.CraftingService>();
-                    var critDef = crafting != null ? crafting.GetStatByMapping(TimelessEchoes.Gear.HeroStatMapping.CritChance) : null;
-                    if (critDef != null)
-                    {
-                        var raw = equip.GetCritChance(critDef);
-                        critChance = critDef.isPercent ? raw / 100f : raw;
-                    }
-                }
-
-                bool isCritical = false;
-                if (critChance > 0f && Random.value < Mathf.Clamp01(critChance))
-                {
-                    total *= 2f;
-                    isCritical = true;
-                }
-
-                var bonusDamage = total - dmgBase;
-                proj.Init(target, total, true, null, combatSkill, bonusDamage, isCritical);
-            }
-        }
-
         private enum State
         {
             Idle,
             MovingToTask,
             PerformingTask,
             Combat
-        }
-
-        public void SetActiveState(bool active)
-        {
-            if (ai != null) ai.enabled = active;
-            if (setter != null) setter.enabled = active;
-            logicActive = active;
-
-            if (!active && animator != null)
-            {
-                animator.SetFloat("MoveX", 0f);
-                animator.SetFloat("MoveY", 0f);
-                animator.SetFloat("MoveMagnitude", 0f);
-            }
-        }
-
-        #region Pathfinding Helpers
-
-        public void SetDestination(Transform dest)
-        {
-            destinationOverride = false;
-            setter.target = dest;
-            ai?.SearchPath();
-        }
-
-        [Button("Mark Destination Reached")]
-        public void SetDestinationReached()
-        {
-            destinationOverride = true;
-        }
-
-        private bool IsAtDestination(Transform dest)
-        {
-            if (dest == null || ai == null) return false;
-            if (destinationOverride) return true;
-            if (ai.reachedDestination || ai.reachedEndOfPath) return true;
-
-            var threshold = ai.endReachedDistance + 0.1f;
-            return Vector2.Distance(transform.position, dest.position) <= threshold;
-        }
-
-        private void AutoAdvance()
-        {
-            if (IsEcho && Instance != null && Instance != this)
-            {
-                var mainHero = Instance.transform;
-
-                if (setter.target != mainHero)
-                {
-                    setter.target = mainHero;
-                    ai?.SearchPath();
-                }
-
-                ai.canMove = true;
-                return;
-            }
-
-            if (idleWalkTarget == null)
-            {
-                idleWalkTarget = new GameObject("IdleWalkTarget").transform;
-                idleWalkTarget.hideFlags = HideFlags.HideInHierarchy;
-            }
-
-            var pos = transform.position;
-            if (idleWalkTarget.position.x - pos.x < 1f)
-                idleWalkTarget.position = new Vector3(pos.x + idleWalkStep, pos.y, pos.z);
-
-            if (setter.target != idleWalkTarget)
-            {
-                setter.target = idleWalkTarget;
-                ai?.SearchPath();
-            }
-
-            ai.canMove = true;
         }
 
         private void OnAutoBuffChanged()
@@ -1139,6 +437,5 @@ namespace TimelessEchoes.Hero
             }
         }
 
-        #endregion
     }
 }
